Create working simulation with uart input.

+++ particle_sim_lab/particle_sim_lab.sdk/particle_sim/src/simulation.h
@@ -9,16 +9,16 @@
 #define SRC_SIMULATION_H_
 
 struct Particle {
-	int x;
-	int y;
-	int vx;
-	int vy;
+	float x;
+	float y;
+	float vx;
+	float vy;
 };
 
 struct Attractor {
-	int x;
-	int y;
-	int g;
+	float x;
+	float y;
+	float g;
 };
 
 void updateSimulation(struct Particle *particles, struct Attractor *attractors,
